+++ packages/core/src/configuration.ts
@@ -39,7 +39,7 @@ import {
 	TruncationTypes,
 	ToolbarControlTypes,
 	ZoomBarTypes,
-	LegendItemType,
+	LegendItemTypes,
 } from './interfaces';
 import enUSLocaleObject from 'date-fns/locale/en-US/index';
 import { circlePack } from './configuration-non-customizable';
@@ -337,7 +337,7 @@ const bubbleChart: BubbleChartOptions = Tools.merge({}, axisChart, {
 	legend: {
 		additionalItems: [
 			{
-				type: LegendItemType.RADIUS,
+				type: LegendItemTypes.RADIUS,
 				name: 'Radius',
 			},
 		],
@@ -359,6 +359,26 @@ const bulletChart: BulletChartOptions = Tools.merge({}, axisChart, {
 			enabled: false,
 		},
 	},
+	legend: {
+		additionalItems: [
+			{
+				type: LegendItemTypes.AREA,
+				name: 'Poor area',
+			},
+			{
+				type: LegendItemTypes.AREA,
+				name: 'Satisfactory area',
+			},
+			{
+				type: LegendItemTypes.AREA,
+				name: 'Great area',
+			},
+			{
+				type: LegendItemTypes.QUARTILE,
+				name: 'Quartiles',
+			},
+		],
+	}
 } as BulletChartOptions);
 
 /*
@@ -395,6 +415,7 @@ const pieChart: PieChartOptions = Tools.merge({}, chart, {
 			enabled: true,
 		},
 		alignment: Alignments.LEFT,
+		sortFunction: null,
 	},
 } as PieChartOptions);