+++ packages/core/src/interfaces/charts.ts
@@ -319,6 +319,7 @@ export interface PieChartOptions extends BaseChartOptions {
 			enabled?: Boolean;
 		};
 		alignment?: Alignments;
+		sortFunction?: (a: any, b: any) => number;
 	};
 }