+++ packages/core/src/interfaces/enums.ts
@@ -232,7 +232,7 @@ export enum AxisTitleOrientations {
 /**
  * enum of legend item type
  */
-export enum LegendItemType {
+export enum LegendItemTypes {
 	CHECKBOX = 'checkbox',
 	RADIUS = 'radius',
 	AREA = 'area',